Create eth code
This code uses the axios library to make a GET request to the CoinDesk API, which returns the current price of Ethereum in USD. The try block attempts to get the response and then the price in USD. If successful, it logs the current price to the console. If there is an error, it will be caught and logged in the catch block.